+++ backend/http/settings.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"github.com/gtsteffaniak/filebrowser/backend/settings"
-	"github.com/gtsteffaniak/filebrowser/backend/users"
 )
 
 type settingsData struct {
@@ -13,14 +12,12 @@ type settingsData struct {
 	CreateUserDir    bool                  `json:"createUserDir"`
 	UserHomeBasePath string                `json:"userHomeBasePath"`
 	Defaults         settings.UserDefaults `json:"defaults"`
-	Rules            []users.Rule          `json:"rules"`
 	Frontend         settings.Frontend     `json:"frontend"`
-	Commands         map[string][]string   `json:"commands"`
 }
 
 // settingsGetHandler retrieves the current system settings.
 // @Summary Get system settings
-// @Description Returns the current configuration settings for signup, user directories, rules, frontend, and commands.
+// @Description Returns the current configuration settings for signup, user directories, rules, frontend.
 // @Tags Settings
 // @Accept json
 // @Produce json
@@ -32,7 +29,6 @@ func settingsGetHandler(w http.ResponseWriter, r *http.Request, d *requestContex
 		CreateUserDir:    config.Server.CreateUserDir,
 		UserHomeBasePath: config.Server.UserHomeBasePath,
 		Defaults:         config.UserDefaults,
-		Rules:            config.Rules,
 		Frontend:         config.Frontend,
 	}
 
@@ -41,7 +37,7 @@ func settingsGetHandler(w http.ResponseWriter, r *http.Request, d *requestContex
 
 // settingsPutHandler updates the system settings.
 // @Summary Update system settings
-// @Description Updates the system configuration settings for signup, user directories, rules, frontend, and commands.
+// @Description Updates the system configuration settings for signup, user directories, rules, frontend.
 // @Tags Settings
 // @Accept json
 // @Produce json
@@ -60,7 +56,6 @@ func settingsPutHandler(w http.ResponseWriter, r *http.Request, d *requestContex
 	config.Server.CreateUserDir = req.CreateUserDir
 	config.Server.UserHomeBasePath = req.UserHomeBasePath
 	config.UserDefaults = req.Defaults
-	config.Rules = req.Rules
 	config.Frontend = req.Frontend
 	config.Auth.Signup = req.Signup
 	err = store.Settings.Save(config)