+++ backend/http/users.go
@@ -19,14 +19,9 @@ import (
 )
 
 var (
-	NonModifiableFieldsForNonAdmin = []string{"Username", "Scope", "LockPassword", "Perm", "Commands", "Rules"}
+	NonModifiableFieldsForNonAdmin = []string{"Username", "Scope", "LockPassword", "Perm"}
 )
 
-// SortingSettings represents the sorting settings.
-type Sorting struct {
-	By  string `json:"by"`
-	Asc bool   `json:"asc"`
-}
 type UserRequest struct {
 	What  string      `json:"what"`
 	Which []string    `json:"which"`
@@ -100,7 +95,9 @@ func userGetHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (
 	u.Password = ""
 	u.ApiKeys = nil
 	if !d.user.Perm.Admin {
-		u.Scope = ""
+		u.Scopes = map[string]string{
+			"default": u.Scopes["default"],
+		}
 	}
 
 	return renderJSON(w, r, u)
@@ -160,7 +157,7 @@ func usersPostHandler(w http.ResponseWriter, r *http.Request, d *requestContext)
 
 	// Validate the user's scope
 	idx := files.GetIndex("default")
-	_, _, err := idx.GetRealPath(d.user.Scope)
+	_, _, err := idx.GetRealPath(d.user.Scopes["default"])
 	if err != nil {
 		return http.StatusBadRequest, err
 	}
@@ -215,13 +212,6 @@ func userPutHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (
 
 	if givenUserId != d.user.ID && !d.user.Perm.Admin {
 		return http.StatusForbidden, nil
-	}
-
-	// Validate the user's scope
-	idx := files.GetIndex("default")
-	_, _, err := idx.GetRealPath(d.user.Scope)
-	if err != nil {
-		return http.StatusBadRequest, err
 	}
 
 	// Read the JSON body