+++ backend/settings/config_test.go
@@ -22,26 +22,6 @@ func TestInitialize(t *testing.T) {
 	}
 }
 
-func Test_loadConfigFile(t *testing.T) {
-	type args struct {
-		configFile string
-	}
-	tests := []struct {
-		name string
-		args args
-		want []byte
-	}{
-		// TODO: Add test cases.
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got, _ := loadConfigFile(tt.args.configFile); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("loadConfigFile() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func Test_setDefaults(t *testing.T) {
 	tests := []struct {
 		name string