+++ backend/users/users.go
@@ -1,8 +1,6 @@
 package users
 
 import (
-	"regexp"
-
 	jwt "github.com/golang-jwt/jwt/v4"
 )
 
@@ -17,15 +15,10 @@ type AuthToken struct {
 }
 
 type Permissions struct {
-	Api      bool `json:"api"`
-	Admin    bool `json:"admin"`
-	Execute  bool `json:"execute"`
-	Create   bool `json:"create"`
-	Rename   bool `json:"rename"`
-	Modify   bool `json:"modify"`
-	Delete   bool `json:"delete"`
-	Share    bool `json:"share"`
-	Download bool `json:"download"`
+	Api    bool `json:"api"`
+	Admin  bool `json:"admin"`
+	Modify bool `json:"modify"`
+	Share  bool `json:"share"`
 }
 
 // SortingSettings represents the sorting settings.
@@ -42,7 +35,7 @@ type User struct {
 	ID                   uint                 `storm:"id,increment" json:"id"`
 	Username             string               `storm:"unique" json:"username"`
 	Password             string               `json:"password,omitempty"`
-	Scope                string               `json:"scope"`
+	Scopes               map[string]string    `json:"scopes"`
 	Locale               string               `json:"locale"`
 	LockPassword         bool                 `json:"lockPassword"`
 	ViewMode             string               `json:"viewMode"`
@@ -50,7 +43,6 @@ type User struct {
 	Sorting              Sorting              `json:"sorting"`
 	Perm                 Permissions          `json:"perm"`
 	Commands             []string             `json:"commands"`
-	Rules                []Rule               `json:"rules"`
 	ApiKeys              map[string]AuthToken `json:"apiKeys,omitempty"`
 	ShowHidden           bool                 `json:"showHidden"`
 	DateFormat           bool                 `json:"dateFormat"`
@@ -61,39 +53,17 @@ type User struct {
 }
 
 var PublicUser = User{
-	Username:     "publicUser", // temp user not registered
-	Password:     "publicUser", // temp user not registered
-	Scope:        "/does/not/exist",
+	Username: "publicUser", // temp user not registered
+	Password: "publicUser", // temp user not registered
+	Scopes: map[string]string{
+		"default": "",
+	},
 	ViewMode:     "normal",
 	LockPassword: true,
 	Perm: Permissions{
-		Create:   false,
-		Rename:   false,
-		Modify:   false,
-		Delete:   false,
-		Share:    false,
-		Download: true,
-		Admin:    false,
-		Api:      false,
+		Modify: false,
+		Share:  false,
+		Admin:  false,
+		Api:    false,
 	},
-}
-
-// GetRules implements rules.Provider.
-func (u *User) GetRules() []Rule {
-	return u.Rules
-}
-
-// CanExecute checks if an user can execute a specific command.
-func (u *User) CanExecute(command string) bool {
-	if !u.Perm.Execute {
-		return false
-	}
-
-	for _, cmd := range u.Commands {
-		if regexp.MustCompile(cmd).MatchString(command) {
-			return true
-		}
-	}
-
-	return false
 }