+++ frontend/src/views/settings/Profile.vue
@@ -19,24 +19,38 @@
             Always show download icon for quick access
           </p>
           <div v-if="hasOnlyOfficeEnabled">
-          <h3>Disable onlyoffice viewer for certain file extentions</h3>
-            <p>A space separated list of file extensions to disable the only office viewer for. eg <code>.txt .html</code></p>
-            <input
-              class="input input--block"
-              :class="{'invalid-form':!formValidation()}"
-              type="text"
-              placeholder="enter file extentions"
-              id="onlyofficeExt"
-              v-model="disableOnlyOfficeExt"
-            />
-            <button class="button" @click="submitOnlyOfficeChange" >save</button>
+            <h3>Disable onlyoffice viewer for certain file extentions</h3>
+            <p>
+              A space separated list of file extensions to disable the only office viewer
+              for. (eg <code>.txt .html</code>)
+            </p>
+            <div class="onlyoffice-group">
+              <input
+                class="input input--block onlyoffice-form"
+                :class="{ 'invalid-form': !formValidation() }"
+                type="text"
+                placeholder="enter file extentions"
+                id="onlyofficeExt"
+                v-model="disableOnlyOfficeExt"
+              />
+              <button class="button onlyoffice-button" @click="submitOnlyOfficeChange">
+                save
+              </button>
+            </div>
           </div>
 
-
           <h3>Theme Color</h3>
-          <ButtonGroup :buttons="colorChoices" @button-clicked="setColor" :initialActive="color" />
+          <ButtonGroup
+            :buttons="colorChoices"
+            @button-clicked="setColor"
+            :initialActive="color"
+          />
           <h3>{{ $t("settings.language") }}</h3>
-          <Languages class="input input--block" :locale="locale" @update:locale="updateLocale"></Languages>
+          <Languages
+            class="input input--block"
+            :locale="locale"
+            @update:locale="updateLocale"
+          ></Languages>
         </div>
       </form>
     </div>
@@ -121,26 +135,29 @@ export default {
   },
   methods: {
     formValidation() {
+      if (this.disableOnlyOfficeExt == "") {
+        return true;
+      }
       let regex = /^\.\w+(?: \.\w+)*$/;
-      return regex.test(this.disableOnlyOfficeExt)
+      return regex.test(this.disableOnlyOfficeExt);
     },
     submitOnlyOfficeChange(event) {
       if (!this.formValidation()) {
-        notify.showError("Invalid input, does not match requirement.")
-        return
+        notify.showError("Invalid input, does not match requirement.");
+        return;
       }
-      this.updateSettings(event)
+      this.updateSettings(event);
     },
     setColor(string) {
-      this.color = string
-      this.updateSettings()
+      this.color = string;
+      this.updateSettings();
     },
     async updateSettings(event) {
       if (event !== undefined) {
         event.preventDefault();
       }
       if (this.color != "") {
-        document.documentElement.style.setProperty('--primaryColor', this.color);
+        document.documentElement.style.setProperty("--primaryColor", this.color);
       }
       try {
         const data = {
@@ -160,7 +177,7 @@ export default {
           "dateFormat",
           "themeColor",
           "quickDownload",
-          "disableOnlyOfficeExt"
+          "disableOnlyOfficeExt",
         ]);
         mutations.updateCurrentUser(data);
         if (shouldReload) {
@@ -180,6 +197,17 @@ export default {
 </script>
 
 <style>
+.onlyoffice-group {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+}
+.onlyoffice-button {
+  margin-left: 1em;
+}
+.onlyoffice-form {
+  height: 3em;
+}
 .invalid-form {
   border-color: red !important;
 }