+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file. For commit guidelines, please refer to [Standard Version](https://github.com/conventional-changelog/standard-version).
 
+## v0.6.0-beta
+
+ **New Features**:
+
+ **Notes**:
+  - downloads no longer open new window.
+  - user permissions simplified to four permission groups (no config change required):
+    - removed  : create, rename, delete, download
+    - remaining: admin, modify, share, api
+    - `scope` is deprecated, but still supported, applies to default source. if using multiple sources, set defaultUserScope at the source instead.
+  - removed user rules and commands.
+    - commands feature has never been enabled so just removing the references.
+    - rules will come back in a different form (not applied to the user).
+
+ **BugFixes**:
+
+
 ## v0.5.4-beta
 
  **BugFixes**:
@@ -12,7 +29,7 @@ All notable changes to this project will be documented in this file. For commit 
   
 ## v0.5.3-beta
 
- **New Features**
+ **New Features**:
   - onlyoffice disable filetypes for user specified file types. https://github.com/gtsteffaniak/filebrowser/issues/346
 
  **Notes**: